Updates to the usage of the LocationManager.
Assign the delegate for the LocationManager as soon as the MapViewController loads. This is so that if the user changes the authorization of the GPS (i.e. from not authorized to authorized or vice-versa, the LocationManager always has a delegate to which to callback and pass the location to.

Additionally, do not show an alert about location services being disabled if authorization status for the LocationManager has not been determined yet (i.e. this is the first time starting the app and the app is requesting location services from the user).