Add the Magnet login flow to the login view.
This commit removes the login flow that used to take place in the map view. Instead that code was abstracted into a Session model. This model is used by the login view to initiate login to the Magnet servers. Currently, to speed up debug, login username / password are hardcoded inside Session.swift so that you don't have to really specify a username / password to login. You can just click login and it'll login with this hardcoded credentials right away.