Integrate the Magnet API for creating a new user.
Users can now be created. However, for convenience, the login view is still hard-coded to login foobar so that we don't have to login each time we run the app.

Additionally, Configurations.plist was restored back to the original file. It is also now gitignore'd so each developer can use their own.