Add the ability to save credentials in credential storage.
With this, credentials can persist across app loads and subsequently, auto-login a user if they have never logged out. However, if a user does logout, the credential in the credential storage is removed and the next time the app loads, the user is not auto-logged-in.

Only one user's credentials are stored at any given time.