Add labels and code for the profile view.
Now when a user logs in, the User instance for that user is cached off by the User class. Later, when the profile view is desired, that same user instance is used to contact Magnet servers to retrieve any further user details (if necessary). Once all details for a user are available, the profile view is populated.